fix: tailor agent handoff guidance to the client surface
